refactor(honeycrisp): split god factory into focused state modules and reorganize files
Split the 538-line createNotesState into three focused factories (foldersState,
notesState, viewState) with a clean dependency DAG. Extract workspace schema from
client instantiation, co-locate editor utilities with the Editor component, and
consolidate date utils. All component imports updated to use the split state modules.
