+++ apps/honeycrisp/src/lib/components/NoteList.svelte
@@ -5,28 +5,17 @@
 	import ArrowUpDownIcon from '@lucide/svelte/icons/arrow-up-down';
 	import CheckIcon from '@lucide/svelte/icons/check';
 	import PlusIcon from '@lucide/svelte/icons/plus';
-	import { differenceInDays, format, isToday, isYesterday } from 'date-fns';
 	import NoteCard from '$lib/components/NoteCard.svelte';
-	import { notesState } from '$lib/state/notes.svelte';
-	import { parseDateTime } from '$lib/utils/date';
+	import { notesState, viewState } from '$lib/state';
+	import { getDateLabel } from '$lib/utils/date';
 	import type { Note } from '$lib/workspace';
 
 	/** Notes to display — filtered active notes or deleted notes depending on view. */
 	const notes = $derived(
-		notesState.isRecentlyDeletedView
+		viewState.isRecentlyDeletedView
 			? notesState.deletedNotes
-			: notesState.filteredNotes,
+			: viewState.filteredNotes,
 	);
-
-	function getDateLabel(dts: string): string {
-		const date = parseDateTime(dts);
-		if (isToday(date)) return 'Today';
-		if (isYesterday(date)) return 'Yesterday';
-		const daysAgo = differenceInDays(new Date(), date);
-		if (daysAgo <= 7) return 'Previous 7 Days';
-		if (daysAgo <= 30) return 'Previous 30 Days';
-		return format(date, 'MMMM yyyy');
-	}
 
 	const groupedNotes = $derived.by(() => {
 		const pinned = notes
@@ -76,18 +65,18 @@
 		if (flatNoteIds.length === 0) return;
 		e.preventDefault();
 
-		const currentIndex = notesState.selectedNoteId
-			? flatNoteIds.indexOf(notesState.selectedNoteId)
+		const currentIndex = viewState.selectedNoteId
+			? flatNoteIds.indexOf(viewState.selectedNoteId)
 			: -1;
 
 		if (e.key === 'ArrowDown') {
 			const nextIndex =
 				currentIndex < flatNoteIds.length - 1 ? currentIndex + 1 : 0;
-			notesState.selectNote(flatNoteIds[nextIndex]!);
+			viewState.selectNote(flatNoteIds[nextIndex]!);
 		} else {
 			const prevIndex =
 				currentIndex > 0 ? currentIndex - 1 : flatNoteIds.length - 1;
-			notesState.selectNote(flatNoteIds[prevIndex]!);
+			viewState.selectNote(flatNoteIds[prevIndex]!);
 		}
 	}
 </script>
@@ -96,10 +85,10 @@
 <div class="flex h-full flex-col" onkeydown={handleKeydown} tabindex="-1">
 	<div class="flex items-center justify-between border-b px-4 py-3">
 		<div class="flex items-center gap-2">
-			<h2 class="text-sm font-semibold">{notesState.folderName}</h2>
+			<h2 class="text-sm font-semibold">{viewState.folderName}</h2>
 			<span class="text-xs text-muted-foreground">{notes.length}</span>
 		</div>
-		{#if !notesState.isRecentlyDeletedView}
+		{#if !viewState.isRecentlyDeletedView}
 			<div class="flex items-center gap-1">
 				<DropdownMenu.Root>
 					<DropdownMenu.Trigger>
@@ -110,24 +99,28 @@
 						{/snippet}
 					</DropdownMenu.Trigger>
 					<DropdownMenu.Content align="end" class="w-44">
-						<DropdownMenu.Item onclick={() => notesState.setSortBy('dateEdited')}>
-							{#if notesState.sortBy === 'dateEdited'}
+						<DropdownMenu.Item
+							onclick={() => viewState.setSortBy('dateEdited')}
+						>
+							{#if viewState.sortBy === 'dateEdited'}
 								<CheckIcon class="mr-2 size-4" />
 							{:else}
 								<span class="mr-2 size-4"></span>
 							{/if}
 							Date Edited
 						</DropdownMenu.Item>
-						<DropdownMenu.Item onclick={() => notesState.setSortBy('dateCreated')}>
-							{#if notesState.sortBy === 'dateCreated'}
+						<DropdownMenu.Item
+							onclick={() => viewState.setSortBy('dateCreated')}
+						>
+							{#if viewState.sortBy === 'dateCreated'}
 								<CheckIcon class="mr-2 size-4" />
 							{:else}
 								<span class="mr-2 size-4"></span>
 							{/if}
 							Date Created
 						</DropdownMenu.Item>
-						<DropdownMenu.Item onclick={() => notesState.setSortBy('title')}>
-							{#if notesState.sortBy === 'title'}
+						<DropdownMenu.Item onclick={() => viewState.setSortBy('title')}>
+							{#if viewState.sortBy === 'title'}
 								<CheckIcon class="mr-2 size-4" />
 							{:else}
 								<span class="mr-2 size-4"></span>
@@ -154,7 +147,7 @@
 				class="flex h-full items-center justify-center p-8 text-center text-muted-foreground"
 			>
 				<p class="text-sm">
-					{#if notesState.isRecentlyDeletedView}
+					{#if viewState.isRecentlyDeletedView}
 						No deleted notes
 					{:else}
 						No notes yet. Click + to create one.
@@ -169,7 +162,7 @@
 							{group.label}
 						</h3>
 						{#each group.entries as note (note.id)}
-						<NoteCard {note} />
+							<NoteCard {note} />
 						{/each}
 					</div>
 				{/each}
