+++ apps/honeycrisp/src/lib/workspace/index.ts
@@ -1,0 +1,8 @@
+export { default } from './client';
+export {
+	type Folder,
+	FolderId,
+	honeycrisp,
+	type Note,
+	NoteId,
+} from './schema';
